Init MovieService
